requirements.txt
accelerate
aiohttp
anyio
attrs
chromadb
crewai
fastapi
ffmpeg-python
filetype
Flask
flask-cors
huggingface-hub
importlib_metadata
Jinja2
langchain
langchain-community
langchain-core
langchain-openai
llama-index
llama-index-core
llama-index-cli
llama-index-agent-openai
llama-index-embeddings-huggingface
llama-index-embeddings-openai
llama-index-llms-huggingface
llama-index-llms-openai
llama-index-vector-stores-chroma
llama-index-readers-file
llama-index-readers-llama-parse
llama-parse
numpy
openai
opencv-python
pandas
python-dotenv
PyYAML
requests
scikit-learn
sentence-transformers
sounddevice
SpeechRecognition
torch
torchaudio
transformers
tqdm
ultralytics
uvicorn
vosk
werkzeug
